pyproject.toml
[tool.poetry]
name = "tinyarag"
version = "0.1.0"
description = "Adaptative RAG simple example"
authors = ["Pere Pasamonte <perepc@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain = "^0.2.6"
langchain-community = "^0.2.6"
langchain-groq = "^0.1.5"
chromadb = "^0.5.3"
pydotenv = "^0.0.7"
langgraph = "^0.1.3"
pytest = "^8.2.2"
sentence-transformers = "^3.0.1"
unstructured = "^0.14.9"
markdown = "^3.6"
psutil = "^6.0.0"
langchainhub = "^0.1.20"
langchain-huggingface = "^0.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
